merge-changes-to-resolve-conflicts-with-12502
