+++ src/app/features/fss-search/fss-simplified-search/fss-simplified-search.component.html
@@ -1,0 +1,13 @@
+<div>
+    <ukho-dialogue [type]='messageType' *ngIf="displayDialogMessage" aria-live="polite">
+      <h3>{{errorMessageTitle}}</h3>
+      <div>{{errorMessageDescription}}</div>
+   </ukho-dialogue>
+
+    <div class="simplifiedSection">
+      <label for="Search">Search</label> <br>   
+        <ukho-textinput class="simplifiedTextbox" [(ngModel)]="fieldValue"></ukho-textinput>
+        <ukho-button (click)="getSimplifiedSearchResult()">Search</ukho-button>
+    </div>
+    <a class="switchToAdvanced" (click)="searchToSimplifiedSearch()">Advanced Search</a>
+</div>
