+++ src/app/features/fss-search/fss-simplified-search/fss-simplified-search.component.scss
@@ -1,0 +1,20 @@
+.simplifiedSection{
+	.text-input-wrap{
+		height: initial;
+	}
+	.simplifiedTextbox{
+	    width: 40%;
+		float: left;
+		margin-right: 20px;
+	}
+}
+.switchToAdvanced{
+	color: #000;
+	font-weight: 400;
+	cursor: pointer;
+	text-decoration: underline;
+	margin-bottom: 20px;
+    width: 40%;
+    display: block;
+    text-align: right;
+}
